pyproject.toml
[tool.poetry]
name = "msadmin-users"
version = "0.1.0"
description = ""
authors = ["Code-Crafters"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = {version = "5.0.3", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
djangorestframework = {version = "3.14.0", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
django-rest-enumfield = {version = "0.2.0", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
flake8 = {version = "7.0.0", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
mypy = {version = "1.9.0", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
django-stubs = {version = "4.2.7", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
psycopg2 = {version = "2.9.9", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
django-cors-headers = {version = "4.3.1", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
asgiref = {version = "3.7.2", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
django-stubs-ext = {version = "4.2.7", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
mccabe = {version = "0.7.0", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
mypy-extensions = {version = "1.0.0", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
pycodestyle = {version = "2.11.1", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
pyflakes = {version = "3.2.0", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
pytz = {version = "2024.1", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
sqlparse = {version = "0.4.4", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
tomli = {version = "2.0.1", markers = "python_version >= \"3.10\" and python_version < \"3.11\""}
types-pytz = {version = "2024.1.0.20240203", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
types-pyyaml = {version = "6.0.12.20240311", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
typing-extensions = {version = "4.10.0", markers = "python_version >= \"3.10\" and python_version < \"4.0\""}
tzdata = {version = "2024.1", markers = "python_version >= \"3.10\" and python_version < \"4.0\" and sys_platform == \"win32\""}
pyjwt = "^2.8.0"
pika = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"